scripts/requirements.txt
PyYAML>=6.0
pypdf>=4.0
